Made the page cache sweeper a little less aggressive
git-svn-id: http://svn.typosphere.org/typo/trunk@1610 820eb932-12ee-0310-9ca8-eeb645f39767

+++ app/models/page_cache.rb
@@ -1,18 +1,24 @@
 class PageCache < ActiveRecord::Base
-  cattr_accessor :public_path
-  @@public_path = ActionController::Base.page_cache_directory
+  def self.public_path
+    ActionController::Base.page_cache_directory
+  end
 
   def self.sweep_all
     self.zap_pages('index.*', 'articles', 'pages')
   end
 
+  def self.sweep_theme_cache
+    self.zap_pages('images/theme', 'stylesheets/theme', 'javascripts/theme')
+  end
+
   def self.zap_pages(*paths)
+    srcs = paths.inject([]) { |o,v|
+      o + Dir.glob(public_path + "/#{v}")
+    }
+    return true if srcs.empty?
     trash = Dir::tmpdir
     FileUtils.makedirs(trash)
-    srcs = paths.collect { |v|
-      Dir.glob(public_path + "/#{v}")
-    }
-    FileUtils.mv(srcs.flatten, trash, :force => true)
+    FileUtils.mv(srcs, trash, :force => true)
     FileUtils.rm_rf(trash)
   end
 end
